+++ source/Processing.Api/Program.cs
@@ -51,6 +51,7 @@ using Processing.Infrastructure.Configuration.DataAccess.Consumers;
 using Processing.Infrastructure.Configuration.DataAccess.EnergySuppliers;
 using Processing.Infrastructure.Configuration.DataAccess.ProcessManagers;
 using Processing.Infrastructure.Configuration.DomainEventDispatching;
+using Processing.Infrastructure.Configuration.Outbox;
 using Processing.Infrastructure.ContainerExtensions;
 using Processing.Infrastructure.EDI;
 using Processing.Infrastructure.EDI.ChangeOfSupplier;
@@ -61,7 +62,6 @@ using Processing.Infrastructure.EDI.MoveIn;
 using Processing.Infrastructure.Integration.IntegrationEvents.EnergySupplierChange;
 using Processing.Infrastructure.Integration.Notifications;
 using Processing.Infrastructure.InternalCommands;
-using Processing.Infrastructure.Outbox;
 using Processing.Infrastructure.Serialization;
 using Processing.Infrastructure.Transport.Protobuf;
 using Processing.Infrastructure.Transport.Protobuf.Integration;
