+++ source/Processing.Infrastructure/Configuration/Outbox/OutboxMessage.cs
@@ -15,7 +15,7 @@
 using System;
 using NodaTime;
 
-namespace Processing.Infrastructure.Outbox
+namespace Processing.Infrastructure.Configuration.Outbox
 {
     public class OutboxMessage
     {
