+++ source/Processing.Infrastructure/Integration/IntegrationEvents/EnergySupplierChange/PublishWhenEnergySupplierHasChanged.cs
@@ -20,7 +20,7 @@ using MediatR;
 using Processing.Domain.EnergySuppliers;
 using Processing.Domain.MeteringPoints.Events;
 using Processing.Infrastructure.Configuration.DataAccess;
-using Processing.Infrastructure.Outbox;
+using Processing.Infrastructure.Configuration.Outbox;
 
 namespace Processing.Infrastructure.Integration.IntegrationEvents.EnergySupplierChange
 {
