--- source/Processing.Infrastructure/Outbox/OutboxMessageCategory.cs
@@ -1,30 +1,0 @@
-﻿// Copyright 2020 Energinet DataHub A/S
-//
-// Licensed under the Apache License, Version 2.0 (the "License2");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-using Processing.Domain.SeedWork;
-
-namespace Processing.Infrastructure.Outbox
-{
-    public class OutboxMessageCategory : EnumerationType
-    {
-        public static readonly OutboxMessageCategory IntegrationEvent = new OutboxMessageCategory(0, nameof(IntegrationEvent));
-        public static readonly OutboxMessageCategory ActorMessage = new OutboxMessageCategory(1, nameof(ActorMessage));
-        public static readonly OutboxMessageCategory MessageHub = new(2, nameof(MessageHub));
-
-        public OutboxMessageCategory(int id, string name)
-            : base(id, name)
-        {
-        }
-    }
-}
